Remove support for old uuid versions
uuid 0.7.0 was released back in 2018 introducing breaking changes to
API's diesel is using internally. Because we exposed in our public api
we couldn't remove it back then.
The next release will be diesel 2.0, so a breaking change is now
possible. Give the time span since the release of the initial
compatible version I'm in favour of removing support for even older
version, as support for them comes with a lot of code duplication. (As
it turns out this support was already broken in our master branch
because we don't test it anymore...)
