+++ diesel/Cargo.toml
@@ -25,8 +25,7 @@ serde_json = { version = ">=0.8.0, <2.0", optional = true }
 time = { version = "0.1", optional = true }
 url = { version = "2.1.0", optional = true }
 percent-encoding = { version = "2.1.0", optional = true }
-uuid = { version = ">=0.2.0, <0.9.0", optional = true, features = ["use_std"] }
-uuidv07 = { version = ">=0.7.0, <0.9.0", optional = true, package = "uuid"}
+uuid = { version = ">=0.7.0, <0.9.0", optional = true}
 ipnetwork = { version = ">=0.12.2, <0.16.0", optional = true }
 num-bigint = { version = ">=0.1.41, <0.3", optional = true }
 num-traits = { version = "0.2", optional = true }
@@ -42,7 +41,7 @@ quickcheck = "0.4"
 
 [features]
 default = ["with-deprecated", "32-column-tables"]
-extras = ["chrono", "serde_json", "uuidv07", "deprecated-time", "network-address", "numeric", "r2d2"]
+extras = ["chrono", "serde_json", "uuid", "deprecated-time", "network-address", "numeric", "r2d2"]
 unstable = ["diesel_derives/nightly"]
 large-tables = ["32-column-tables"]
 huge-tables = ["64-column-tables"]
