+++ diesel/src/pg/types/mod.rs
@@ -19,8 +19,6 @@ mod ranges;
 mod record;
 #[cfg(feature = "uuid")]
 mod uuid;
-#[cfg(feature = "uuidv07")]
-mod uuid_v0_7;
 
 /// PostgreSQL specific SQL types
 ///
